:art: refactor(paginate): Rename `paginate` to `pagination`
GTC-1865
